requirements.txt
lightrag-hku
neo4j
sentence-transformers
openai
python-dotenv
pytest
faiss-cpu